+++ src/tune/notestream.hpp
@@ -10,19 +10,19 @@ class Loop;
 class NoteStream {
 protected:
     /// Returns the place where t should be in playable
-    int getId(float t);
-    int getSetterId(float t);
+    int getId(double t);
+    int getSetterId(double t);
 
     /// Thre Data
     /// Design decisions:
     ///     This needs to be performant. That can be easily achieved by placing
     ///     the notes close to eachother. This way of implementing achieves that.
     ///     On the other hand this would be way more pleasing if 
-    std::vector<std::pair<float, Note>> playable;
-    std::vector<std::pair<float, SetterNote>> setter;
+    std::vector<std::pair<double, Note>> playable;
+    std::vector<std::pair<double, SetterNote>> setter;
 
     /// Beats per minute, used for parsing (so you don't need to count the seconds)
-    float bpm = 60;
+    double bpm = 60;
 
     /// Enables playing multiple notes at the same time.
     /// This is achieved by changing the parsing format to include a timestamp 
@@ -31,20 +31,20 @@ protected:
     
 public:
     /// self explanatory getters/setters
-    inline void setBpm(float val) { bpm = val; }
-    inline float getBpm() const { return bpm;}
+    inline void setBpm(double val) { bpm = val; }
+    inline double getBpm() const { return bpm;}
     inline void setPolynote(bool val) { polynote = val; }
     inline bool getPolynote() const { return polynote; }
 
     /// Returns how many seconds it takes to play the stream
-    inline float getLen() {
+    inline double getLen() {
         if(playable.size() == 0) return 0;
-        std::pair<float, Note> last = playable[playable.size()-1];
+        std::pair<double, Note> last = playable[playable.size()-1];
         return last.first + last.second.getLen();
     }
 
-    inline float getPlayableSize() { return playable.size(); }
-    inline float getSetterSize() { return setter.size(); }
+    inline double getPlayableSize() { return playable.size(); }
+    inline double getSetterSize() { return setter.size(); }
 
     inline SetterNote& getSetterNote(int id) {
         if(id < 0 || id >= setter.size())
@@ -61,18 +61,18 @@ public:
     /// constructors
     inline NoteStream() : playable(), setter() {}
     inline NoteStream(Note n);
-    inline NoteStream(float t, Note n);
-    inline NoteStream(std::pair<float, Note> playable);
+    inline NoteStream(double t, Note n);
+    inline NoteStream(std::pair<double, Note> playable);
     NoteStream(const NoteStream& s);
  
     /// Adds a note at the appropriate location
-    void Add(std::pair<float, Note> p);
-    void Add(std::pair<float, SetterNote> p);
-    void Add(std::pair<float, Loop> p);
+    void Add(std::pair<double, Note> p);
+    void Add(std::pair<double, SetterNote> p);
+    void Add(std::pair<double, Loop> p);
 
     /// Returns all notes that start before t and have not been started yet
-    virtual std::vector<Note> GetStartingPlayableNotes(float t);
-    virtual std::vector<SetterNote> GetStartingSetterNotes(float t);
+    virtual std::vector<Note> GetStartingPlayableNotes(double t);
+    virtual std::vector<SetterNote> GetStartingSetterNotes(double t);
 };
 
 /// Parser
